+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Unreleased Changes
 ------------------
 
+* Issue - Aws::S3::Presigner - Added validation to `#presigned_url` to
+  ensure the `:key` option is not an empty string. This prevents a
+  presigned `:get_object` URL from becoming a presigned `:list_objects`
+  request on the bucket.
+
 2.1.19 (2015-09-03)
 ------------------